requirements.txt
fastapi
uvicorn[standard]
litellm
chromadb
pyyaml
gitpython
requests
websockets
httpx
